// src/api/model/taxFilingResponseDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * accounting
 * Accounting API description
 * OpenAPI spec version: 1.0
 */
import type { TaxFilingResponseDtoStatus } from './taxFilingResponseDtoStatus';
import type { TaxFilingResponseDtoSubmissionType } from './taxFilingResponseDtoSubmissionType';
import type { TaxFilingResponseDtoSummary } from './taxFilingResponseDtoSummary';
import type { TaxFilingInvoiceRefDto } from './taxFilingInvoiceRefDto';

export interface TaxFilingResponseDto {
  id: string;
  companyId: string;
  month: number;
  year: number;
  status: TaxFilingResponseDtoStatus;
  submissionType: TaxFilingResponseDtoSubmissionType;
  createDate: string;
  /** @nullable */
  submissionDate?: string | null;
  /** @nullable */
  summary?: TaxFilingResponseDtoSummary;
  /** @nullable */
  errorMessage?: string | null;
  invoices: TaxFilingInvoiceRefDto[];
  createdAt: string;
  updatedAt: string;
}